requirements.txt
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
psycopg2==2.8.4
pycparser==2.19
PyJWT==1.7.1
python-dotenv==0.10.5
requests==2.22.0
six==1.14.0
SQLAlchemy==1.3.12
urllib3==1.25.8
Werkzeug==0.16.0
